vga: add xor-pattern rasterizer.
This is probably the easiest procedural texture to implement, which is
why it's a staple of demo folks.  And while the black-white striped
calibration pattern is a better test of horizontal resolution (and thus
pixel timing), this makes jitter a lot more obvious!  In practice, the
driver usually jitters in units of four pixels (the size of the DMA
fifo), so any pattern that repeats inside that space hides jitter.

With this pattern displayed, it's clear that we have pretty nasty
jitter!  Time to fix that.