demo/hires_graphics: fill instead of undraw.
My line-drawing implementation is now reasonably fast, but the model
drawing code still involves tens of thousands of matrix multiplies.
As a result, it turns out to be way faster to just fill-rect the
framebuffer than to undraw the lines.  Since even an impossibly
optimized line-drawing routine will cap out at 1/32 the bandwidth of
a mediocre fill-rect routine (due to bitbanding), this makes sense.