Enable debugging during WFI.
By default, the STM32 powers down the AHB that drives the JTAG unit
during even the highest of low-power modes -- the one accessed by
WFI by default.  Normally I don't notice this, but occasionally I
have a bug that causes the code to loop in WFI...and this is my least
favorite time to lose the ability to debug or flash the chip.