demo/rook: improve constexpr-ness of matrix ops.
I noticed that the chain of matrix multiplies during setup was being
evaluated at runtime, when I had intended it to happen at compile time.
On inspection, I'd forgotten to constexpr the Matrix operators and
transpose.

Appropriate application of constexpr causes the matrix/vector ops to
finally inline, meaning that matrices are being held in the FPU instead
of spilled to memory between uses.  This cuts rendering time by 5.8ms
per frame, or up to 21%.