demo/rook: remove packed attribute.
So it turns out that GCC (4.6.3) will generate conservative access
sequences for integers inside of structs marked with the packed
attribute.  This is both bad and wrong: bad because it costs several
cycles, and wrong because this platform supports unaligned access to
integers, but not floats, and yet GCC was not being conservative when
loading the float fields.

In practice, because I stopped using binary STL, my data is always
aligned.  I've removed the packed attribute, which means I'm
technically relying on knowledge of struct layout in the AEABI...
but that's not the first ARMv7M-specific thing I'm relying on, so
it won't kill me.  And it makes the rook demo 1.5% faster.