demo/i2s: basic I2S/I2C support for audio codec.
This plays a little tune out the audio codec on the discovery board.
Unfortunately, I've written it in such a way that it conflicts with
video; I use the VGA code to set up clocks, but then disable it.

The problem is that the codec *requires* samples to be available in
a timely fashion, or it generates noise.  Since I'm feeding it in
software, I've essentially created another latency-sensitive task,
but without the scheduling analysis required to actually support
that.  I can fix this pretty trivially by feeding the I2S unit through
a buffer served by DMA, but getting it working as a hack was more
fun for vacation.