Redesign init process to work with generic crt.
We were using a custom crt because we'd like to establish some
processor state invariants before any static constructors run:

 - Floating point is on.
 - Faults are on (to improve diagnostics).
 - Memory is remapped so SRAM is accessible on I/D ports.
 - Initialized data (and RAM-resident code) are initialized.
 - BSS is cleared.

The latter two can be handled by any CRT if we describe the
process carefully enough.  This change rearranges the linker
script and friends so that we can potentially switch to the
ETL CRT.

The first three are only special if we require them to happen
*before* SRAM initialization.  But we don't really need to do
that, if the CRT memory-shape symbols point to the pre-remap
locations.  So I've also changed that.

Now our particular initialization guarantees can be met using
functions in the preinit_array.