Initial Cobble support.
The BUILD files are mechanical translations of the Makefiles, with a few
errors fixed.  (Cobble is significantly stricter about dependencies.)
I'll make another pass to optimize them.

This version of the Cobble build is 6.2x faster than the Make build on
my i5-3570K.  The STL compiler now accounts for a significant fraction
of the build time -- though unlike the Make build, it's only run when
needed.