Simplify Cobble build.
I've centralized the linker configuration for the standalone demos (i.e.
those that don't use toolchain resources).  Which is good, because the
demos varied in how they set up such an environment, and some were doing
it wrong.

I was also able to extend this environment to demo/blt.

Now no targets have to specify extra deltas, and the local deltas are
pretty minimal -- typically having to do with optimization levels, which
probably need tweaking anyhow.