demo/tunnel: substantially rewrite, rearrange.
I recently noticed that (1) __fp16 is actually implemented on ARMv7-M,
at least as of GCC 4.8, and (2) libstdc++'s trig operations are indeed
constexpr.

This change:

- Uses GCC's __fp16 type where possible.
- Factors table generation out of demo main routine.
- Allows compile-time table generation.

The lookup table is always in RAM; when using a compile-time generated
table, it gets copied to RAM to ensure zero-wait-state access.  The two
modes can be switched by editing the TABLE_IN_ROM #define in tunnel.cc.

In practice, this is mostly interesting because when using a ROM table,
no trig routines are linked into the binary.