demo/tunnel: optimize inner loop.
The previous version was written for clarity rather than performance,
and did a lot of unnecessary work -- looking up the same table entry
four times, for example, once for each quadrant.

I've rewritten it as a loop over a single quadrant, generating four
pixels per iteration.  Conveniently, this recognizes that diagonally
opposed quadrants are identical and reuses the shader output, cutting
the cost of shader effects -- so I've added some.