demo/tunnel: use table in-place in ROM.
(This affects the table-in-ROM variant only.)

Previously, the ROM table was copied into SRAM before execution.  I did
this because the previous tunnel renderer produced pixels in raster
order, meaning it produced table lookups out of order -- so we rarely
hit the Flash cache.  Directly accessing the table in ROM meant we
couldn't quite keep up with rendering.

The new renderer accesses the table in-order and produces pixels
out-of-order.  Now the Flash cache can do its job.

As a result we save 30,000 bytes of RAM... just about enough for a
framebuffer....