demo/tunnel: halve framebuffers.
Remember how we're only having to store a quadrant's worth of trig and
reciprocal lookup tables, because the demo is radially symmetric?

While the rendered output is not *radially* symmetric because of the
quadrant shifts, it *is* laterally symmetric.  I've exploited that here
by rendering only the top half of the screen, and using a modified
rasterizer to mirror it onto the bottom.

This doesn't cut rendering *time* by half, exactly -- the rendering was
already taking advantage of the symmetry -- but it *does* cut our RAM
requirement in half.