demo/tunnel: use Glitch tech to quadruple resolution.
I'm backporting more stuff I developed for Glitch -- in this case, an
efficient linear-interpolated table lookup routine that lets us pump
four times the pixels for less CPU cost.

If Glitch's tunnel looked prettier than the demo tunnel here... well,
this is most of the reason why.

Crazily, despite having the pregenerated table as part of the ROM image,
this version is *dramatically smaller* -- 13k instead of 35k -- because
it doesn't have to link in any of GCC's code for atan2/sqrt.