runtime/startup: fix subtle and stupid mistake.
This is a good one.

The original startup code is some of the oldest code in this repo, and
I've massaged and rearranged it as the system grew around it.  Turns out
that one of those massages made things worse.

A bit over a year ago, in commits fa44ae53 and 117c3d0a, I rearranged
the startup routine to match ETL's CRT startup routine, so that I could
quit maintaining two of them.  This involved renaming a whole lot of
symbols used to designate memory areas to more closely match what they
were called 40 years ago on early Unix.  For instance, the family of
symbols called _sram_image_* became _data_*.

One of the places these symbols were used was in setting the address of
the interrupt vector table, using the ARM VTOR register.  So I had to
update that, and I did it wrong.

The wrongness is pretty subtle -- I haven't noticed it for over a year.
Here's the deal:

- Instead of configuring the vector table at (remapped SRAM address)
  0x0000_0000, I configured it at its pre-remap address, 0x2000_0000.

- This is not a functional difference, because the two addresses alias
  the same 112kiB SRAM bank, so it worked fine.

- But the *reason* we're remapping SRAM is that, when SRAM is mapped at
  low addresses, the CPU can access it using its I and D ports.  At the
  SRAM's native high address, it can only be accessed using the S port.

- If the vector table is accessible through the I port, the CPU will
  load the vector address concurrently with the register save on the D
  port.  If not, concurrent access appears to be disabled, even if it's
  theoretically possible (since the S and D ports are independent).

I say "appears to be" because this isn't precisely specified anywhere I
can find.

Empirically, fixing the VTOR address to point to the SRAM alias
accessible through the I/D ports improves throughput very slightly -- on
the order of one cycle per interrupt taken.

Plus, I got to delete some code.