math/conversion: fix typo that broke ceil.
When transcribing and documenting this file for publication, I appear to
have fat-fingered the most important instruction in this assembly
sequence.

The net effect: both ceil and floor acted as round.  For floor, that's
hard to notice (with positive numbers).  With ceil... it was pretty
apparent once I started using it.