demo/raycast: sketch of a raycaster.
This is based heavily on Lode Vandevenne's tutorial, though I've already
made some performance improvements to keep it to 60fps.  (Lode and I are
using very different computers with very different operation costs.)

I'm using a trick to increase the apparent vertical resolution.  This
trick will make it more difficult to do sprites.  If that's a problem I
can reduce the resolution and use a full framebuffer.